User Resolvers
* Updates the users resolver.
  * Updates to the resolver to not return null.
  * Removes the console logging.
* Updates the user resolver
  * Updates the resolver to return an error message when the user does not exist.
  * Removes the console logging.
* Updates the createUser resolver.
  * Validates the password strength.
  * Removes the console logging.
* Updates the updateUser resolver.
  * Updates the resolver to return an error message when the user does not exist.
  * Removes the console logging.
* Updates the deleteUser resolver.
  * Removes the console logging.

User Schema
* Adds email and user name properties.
* Separates UpsertUser into CreateUser and UpdateUser.

+++ src/database/mongoose/mongoose.users.ts
@@ -1,4 +1,7 @@
+import mongoose from 'mongoose';
+
 import { IUsersDatabase } from '../interfaces/database.users';
+import { InvalidDatabaseIdError } from '../interfaces/errors';
 import { CreateUser, UpdateUser, User } from '../models/user';
 
 import { UserDocument2User } from './helpers/mappers';
@@ -8,61 +11,82 @@ class Users implements IUsersDatabase {
 	public async getUsers(): Promise<User[]> {
 		try {
 			const userDocuments: IUserDocument[] = await UserModel.find();
-			if (Array.isArray(userDocuments)) {
-				const users: User[] = new Array();
-				userDocuments.forEach((document: IUserDocument) => {
-					const user = UserDocument2User(document);
-					if (user) {
-						users.push(user);
-					}
-				});
+			const users: User[] = userDocuments.map((doc: IUserDocument) => {
+				return UserDocument2User(doc);
+			});
 
-				return users;
+			return users;
+		} catch (err: any) {
+			console.error(err); // TODO: transfer to a logger
+			throw new Error('Error retrieving users.');
+		}
+	}
+	public async getUserById(_id: string): Promise<User | null> {
+		if (!mongoose.isValidObjectId(_id)) {
+			throw new InvalidDatabaseIdError('Invalid User ID.');
+		}
+		try {
+			const userDocument: IUserDocument | null = await UserModel.findOne({ _id });
+			if (userDocument) {
+				return UserDocument2User(userDocument);
 			}
 		} catch (err: any) {
 			console.error(err); // TODO: transfer to a logger
-			throw new Error('Error occurred getting users.');
-		}
-
-		return new Array();
-	}
-	public async getUserById(_id: string): Promise<User | null> {
-		const userDocument: IUserDocument | null = await UserModel.findOne({ _id });
-		if (userDocument) {
-			return UserDocument2User(userDocument);
+			new Error(`Error retrieving user by ID: ${_id}.`);
 		}
 
 		return null;
 	}
 	public async getUserByUsername(username: string): Promise<User | null> {
-		const userDocument: IUserDocument | null = await UserModel.findOne({ username });
-		if (userDocument) {
-			return UserDocument2User(userDocument);
+		try {
+			const userDocument: IUserDocument | null = await UserModel.findOne({ username });
+			if (userDocument) {
+				return UserDocument2User(userDocument);
+			}
+		} catch (err: any) {
+			console.error(err); // TODO: transfer to a logger
+			throw new Error(`Error retrieving user by username: ${username}`);
 		}
 
 		return null;
 	}
 	public async createUser(user: CreateUser): Promise<User | null> {
-		const userDocument: IUserDocument | void = await UserModel.create({ ...user });
-		if (userDocument) {
-			return UserDocument2User(userDocument);
+		try {
+			const userDocument: IUserDocument = await UserModel.create({ ...user });
+			if (userDocument) {
+				return UserDocument2User(userDocument);
+			}
+		} catch (err: any) {
+			console.error(err); // TODO: transfer to a logger
+			throw new Error('Error creating user.');
 		}
 
 		return null;
 	}
 	public async updateUserById(_id: string, user: UpdateUser): Promise<User | null> {
-		const currentUser: User | null = await this.getUserById(_id);
-		if (!currentUser) {
-			return null;
+		if (!mongoose.isValidObjectId(_id)) {
+			throw new InvalidDatabaseIdError('Invalid User ID.');
 		}
-		const updateResult = await UserModel.updateOne({ _id }, user);
-		if (updateResult.matchedCount == 1) {
-			return { ...currentUser, ...user };
+		try {
+			const currentUser: User | null = await this.getUserById(_id);
+			if (!currentUser) {
+				return null;
+			}
+			const updateResult = await UserModel.updateOne({ _id }, user);
+			if (updateResult.matchedCount == 1) {
+				return { ...currentUser, ...user };
+			}
+		} catch (err: any) {
+			console.error(err); // TODO: transfer to a logger
+			throw new Error('Error updating user.');
 		}
 
 		return null;
 	}
 	public async updatePasswordByUserId(_id: string, password: string): Promise<boolean> {
+		if (!mongoose.isValidObjectId(_id)) {
+			throw new InvalidDatabaseIdError('Invalid User ID.');
+		}
 		try {
 			const updateResult = await UserModel.updateOne({ _id }, { password });
 			if (updateResult) {
@@ -70,13 +94,21 @@ class Users implements IUsersDatabase {
 			}
 		} catch (err: any) {
 			console.error(err); // TODO: transfer to a logger
-			throw new Error(err.message);
+			throw new Error('Error updating user password.');
 		}
 
 		return false;
 	}
 	public async deleteUserById(_id: string): Promise<void> {
-		await UserModel.deleteOne({ _id });
+		if (!mongoose.isValidObjectId(_id)) {
+			throw new InvalidDatabaseIdError('Invalid User ID.');
+		}
+		try {
+			await UserModel.deleteOne({ _id });
+		} catch (err: any) {
+			console.error(err); // TODO: transfer to a logger
+			throw new Error('Error deleting user.');
+		}
 
 		return;
 	}
